merge config.yaml into conanfile.py
It is very difficult to make conanfile.py consist of multiple files due
to the constantly changing path. Until such time that we figure out a
better way, integrate the yaml config as a string in the main
conanfile.py and use that as the single source of truth for
configuration and markdown table generation.